finetune-ERP-backend-New/requirements-test.txt
pytest==8.4.1
pytest-django==4.11.1
freezegun==1.5.5
black==24.4.2
